docs: refactor `lib` folder, add `chai` for shallow object testing (#9)
This PR refactors the `mdx/lib` folder. The directory is now as the following:

```
lib
├── filesystem.ts
├── json.test.ts
├── json.ts
├── mdx.test.ts
├── mdx.ts
├── messages.ts
├── metadata.ts
├── packages.ts
├── record.ts
├── services.ts
├── test-resources
│   ├── booking-messages.mdx
│   ├── booking-services.mdx
│   └── location-messages.mdx
└── types.ts
```

This PR also adds `chai` for a more rich testing experience, for example, the built-in `assert.equal` only tests deep-equality. In `chai`, we can use `.eql` to test shallow equality.

Some function APIs also were changed. For example:

```ts
// Previous.
getMessageString({
  message,
  innerMessagesRecord,
  packageName: file.package,
  level: 1,
})

// After.
getMessageDefinition({
  header: getMessageHeader(message.name),
  body:
    description +
    getMessageProtosaurusBlock({
      packageName: file.package,
      level: 1,
      message,
      innerMessagesRecord,
    }),
})
```

The intention of the API change was so that it is less "drill-down" type of functions. Previously, we are drilling down like 3-4 functions deep, now we are just drilling 1-2 functions deep. The less functions we need to drill down, the faster we can navigate through a function.

Signed-off-by: Try Ajitiono <ballinst@gmail.com>