Add DEVELOPER.md (#18)
This change adds `DEVELOPER.md` to the root of the project. It contains the step-by-step on how the modules work together, as well as how to develop and quickly test them locally.

This PR also adds a Node version requirement (I put 16 but I'm open for a change) to the `mdx` and `website` folders. This is done by adding `engines.node` field in `package.json`.

Signed-off-by: Try Ajitiono <ballinst@gmail.com>