cli: move the MDX/JSON dictionary generation process to `@protosaurus/cli` (#29)
This PR moves the MDX/JSON dictionary generation process to `@protosaurus/cli`. The previously asynchronous to synchronous functions were reverted, so now they are all now asynchronous again (the functions from `fs-extra`).

Upon `prebuild`, the builder will generate the protoc-gen-doc JSON, continued with the MDX/JSON dictionary generation, with one condition: the list of files have changed. In the future, this probably better be a hash of file contents instead of just list of file names.

Signed-off-by: Try Ajitiono <ballinst@gmail.com>