feat: add line highlight feature for `protosaurus` code language (#31)
This PR implements the following syntax:

1. `// highlight-next-line`
2. `// highlight-start` and `// highlight-end`.

This PR also reworks how the deletion works for the MDX files. Instead of whitelisting a static variable, we now will delete files based on the generated JSON files, e.g. if `booking.v1.json` exists, then `booking.v1.mdx` will be deleted. This is handy so that we only need to update `.gitignore` when creating a new static MDX file instead of having to update both `.gitignore` and the `emit.ts` file.

Signed-off-by: Try Ajitiono <ballinst@gmail.com>