plugin: implement collapsible code block for other languages (#34)
This PR implements collapsible code block for other languages. Due to the nature of the metastring for normal code block, the plugin now "trims" the properties that are related to code title. This is done so that Prism.js will not duplicate the code title. The steps are as the following:

1. Parse the metastring
2. Wrap the code element using our `wrapWithMetastringElements` function
3. Trim the strings and keys that are related to the code title

Signed-off-by: Try Ajitiono <ballinst@gmail.com>