plugin: implement line highlighting metastring for protosaurus language  (#35)
This PR implements line highlighting metastring for protosaurus language. Hence, we can do this:

````
```protosaurus collapsible title="Hello world" {1,3,5-8}
message Booking {
  // ...
}
```
````

Technical wise, we have one new information in `MetastringInfo`, `highlightedLines`. This field is the one that we will use to highlight lines. When we iterate lines in the code block, then it will be used to determine whether the line should be highlighted or not.

```ts
export interface MetastringInfo {
  title: string;
  isCollapsible: boolean;
  highlightedLines: number[];
}
```

Signed-off-by: Try Ajitiono <ballinst@gmail.com>